# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FaultTrace-RAG API",
    "description": "Counterfactual Fault Localization for Corpus-Level LLM Analytics Pipelines. Provides deterministic corpus data, procedural queries, dual gold evaluation, and traced pipeline execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/worlds": {
      "get": {
        "tags": [
          "Worlds"
        ],
        "summary": "List all corpus worlds",
        "operationId": "list_worlds_api_v1_worlds_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WorldResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worlds/{world_id}": {
      "get": {
        "tags": [
          "Worlds"
        ],
        "summary": "Get world details",
        "operationId": "get_world_api_v1_worlds__world_id__get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worlds/{world_id}/records": {
      "get": {
        "tags": [
          "Worlds"
        ],
        "summary": "List records in a world with pagination and filters",
        "operationId": "list_world_records_api_v1_worlds__world_id__records_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 5.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rating"
            }
          },
          {
            "name": "max_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 5.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Rating"
            }
          },
          {
            "name": "verified_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Verified Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "PaginatedResponse"
      },
...
      "PaginatedResponse_WorldResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorldResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "PaginatedResponse[WorldResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorldResponse": {
        "properties": {
          "world_id": {
            "type": "string",
            "title": "World Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "scale_n": {
            "type": "integer",
            "title": "Scale N"
          },
          "parent_world_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent World Id"
          },
          "creation_policy": {
            "type": "string",
            "title": "Creation Policy"
          },
          "record_ids_hash": {
            "type": "string",
            "title": "Record Ids Hash"
          },
          "manifest_path": {
            "type": "string",
            "title": "Manifest Path"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "world_id",
          "dataset_id",
          "seed",
          "scale_n",
          "creation_policy",
          "record_ids_hash",
          "manifest_path",
          "created_at",
          "schema_version"
        ],
        "title": "WorldResponse"
      }
    }
  }
}